# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AETHEX Medical Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number;
  category: string;
  categorySlug: string;
  brand: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  tags: string[];
  featured: boolean;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  productCount: number;
  iconName: string;
}

export interface CartItem {
  id: number;
  productId: number;
  name: string;
  price: number;
  imageUrl: string;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartRequest {
  productId: number;
  quantity?: number;
  sessionId: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiChatRequest {
  message: string;
  conversationHistory?: ChatMessage[];
}

export interface AiChatResponse {
  message: string;
  suggestions?: string[];
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetCartParams = {
  sessionId: string;
};

export type RemoveFromCartParams = {
  sessionId: string;
};