# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AETHEX Medical Store API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List products
 */
export const listProductsQueryPageDefault = 1;
export const listProductsQueryLimitDefault = 20;

export const ListProductsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
  page: zod.coerce.number().default(listProductsQueryPageDefault),
  limit: zod.coerce.number().default(listProductsQueryLimitDefault),
});

export const ListProductsResponse = zod.object({
  products: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      description: zod.string(),
      price: zod.number(),
      originalPrice: zod.number().optional(),
      category: zod.string(),
      categorySlug: zod.string(),
      brand: zod.string(),
      imageUrl: zod.string(),
      rating: zod.number(),
      reviewCount: zod.number(),
      inStock: zod.boolean(),
      tags: zod.array(zod.string()),
      featured: zod.boolean(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Get product by ID
 */
export const GetProductParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProductResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  price: zod.number(),
  originalPrice: zod.number().optional(),
  category: zod.string(),
  categorySlug: zod.string(),
  brand: zod.string(),
  imageUrl: zod.string(),
  rating: zod.number(),
  reviewCount: zod.number(),
  inStock: zod.boolean(),
  tags: zod.array(zod.string()),
  featured: zod.boolean(),
});

/**
 * @summary List all categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  slug: zod.string(),
  description: zod.string(),
  productCount: zod.number(),
  iconName: zod.string(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get cart items
 */
export const GetCartQueryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const GetCartResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      name: zod.string(),
      price: zod.number(),
      imageUrl: zod.string(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Add item to cart
 */
export const addToCartBodyQuantityDefault = 1;

export const AddToCartBody = zod.object({
  productId: zod.number(),
  quantity: zod.number().default(addToCartBodyQuantityDefault),
  sessionId: zod.string(),
});

export const AddToCartResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      name: zod.string(),
      price: zod.number(),
      imageUrl: zod.string(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Remove item from cart
 */
export const RemoveFromCartParams = zod.object({
  itemId: zod.coerce.number(),
});

export const RemoveFromCartQueryParams = zod.object({
  sessionId: zod.coerce.string(),
});

export const RemoveFromCartResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      productId: zod.number(),
      name: zod.string(),
      price: zod.number(),
      imageUrl: zod.string(),
      quantity: zod.number(),
    }),
  ),
  total: zod.number(),
  itemCount: zod.number(),
});

/**
 * @summary Chat with AI medical assistant
 */
export const AiChatBody = zod.object({
  message: zod.string(),
  conversationHistory: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod.string(),
      }),
    )
    .optional(),
});

export const AiChatResponse = zod.object({
  message: zod.string(),
  suggestions: zod.array(zod.string()).optional(),
});