# lib/api-zod/src/generated/types/aiChatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AETHEX Medical Store API
 * OpenAPI spec version: 0.1.0
 */

export interface AiChatResponse {
  message: string;
  suggestions?: string[];
}