# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AETHEX Medical Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number;
  category: string;
  categorySlug: string;
  brand: string;
  imageUrl: string;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  tags: string[];
  featured: boolean;
}