# lib/api-zod/src/generated/types/productListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AETHEX Medical Store API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}